feat(pipeline): add debug flag

+++ func/execution/rv64i.py
@@ -370,7 +370,7 @@ def SRA(data):
     shamt = reg_type(data["read_regs"]["int"][1]["value"]) & reg_type(0x3F)
     data["write_regs"]["int"][0]["value"] = reg_type(
         (reg_type(data["read_regs"]["int"][0]["value"]) >> shamt)
-        | reg_type(int(str(sign) * shamt + "0" * (reg_type(0).itemsize * 8 - shamt), 2))
+        | reg_type(int(str(sign) * shamt + "0" * int(reg_type(0).itemsize * 8 - shamt), 2))
     )
     return data
 
@@ -606,7 +606,7 @@ def SRAIW(data):
         sext_word_type(
             (word_type(data["read_regs"]["int"][0]["value"]) >> shamt)
             | word_type(
-                int(str(sign) * shamt + "0" * (word_type(0).itemsize * 8 - shamt), 2)
+                int(str(sign) * shamt + "0" * int(word_type(0).itemsize * 8 - shamt), 2)
             )
         )
     )
@@ -660,7 +660,7 @@ def SRAW(data):
         sext_word_type(
             (word_type(data["read_regs"]["int"][0]["value"]) >> shamt)
             | word_type(
-                int(str(sign) * shamt + "0" * (word_type(0).itemsize * 8 - shamt), 2)
+                int(str(sign) * shamt + "0" * int(word_type(0).itemsize * 8 - shamt), 2)
             )
         )
     )
